// schemas/server/docs/api/ref/api.ts
export interface paths {
  "/api/v2/product/{barcode}": {
    /**
     * Get information for a specific product by barcode 
     * @description A product can be fetched via its unique barcode.
     * It returns all the details of that product response.
     */
    get: operations["get-product-by-barcode"];
  };
  "/api/v2/product/{barcode}?fields=knowledge_panels": {
    /**
     * Get Knowledge panels for a specific product by barcode
     * (special case of get product)
     *  
     * @description Knowledge panels gives high leve informations about a product,
     * ready to display.
     * This is used by open food facts website,
     * and by the official mobile application
     */
    get: operations["get-product-by-barcode-knowledge-panels"];
  };
  "/cgi/product_image_upload.pl": {
    /**
     * Add a Photo to an Existing Product 
     * @description Photos are source and proof of data.
     * The first photo uploaded for a product is
     * auto-selected as the product’s “front” photo.'
     */
    post: operations["get-cgi-product_image_upload.pl"];
  };
  "/cgi/ingredients.pl": {
    /**
     * Performing OCR on a Product 
     * @description Open Food Facts uses optical character recognition (OCR) to retrieve nutritional data and other information from the product labels.
     */
    get: operations["get-cgi-ingredients.pl"];
  };
  "/cgi/product_image_crop.pl": {
    /**
     * Rotate A Photo 
     * @description Although we recommend rotating photos manually and uploading a new version of the image,
     * the OFF API allows you to make api calls to automate this process.
     * You can rotate existing photos by setting the angle to 90º, 180º, or 270º clockwise.
     */
    get: operations["get-cgi-product_image_crop.pl"];
    /**
     * Crop A Photo 
     * @description Cropping is only relevant for editing existing products.
     * You cannot crop an image the first time you upload it to the system.
     */
    post: operations["post-cgi-product_image_crop.pl"];
  };
  "/cgi/product_image_unselect.pl": {
    /** Unselect A Photo */
    post: {
      requestBody?: {
        content: {
          "multipart/form-data": external["requestBodies/unselect_a_photo.yaml"];
        };
      };
      responses: {
        /** @description OK */
        200: {
          content: {
            "application/json": {
              /**
               * @description status of the unselect operation 
               * @example status ok
               */
              status?: string;
              /**
               * @description status code of the operation 
               * @example 0
               */
              status_code?: number;
              /**
               * @description image field that was unselected 
               * @example front_fr
               */
              imagefield?: string;
            };
          };
        };
      };
    };
  };
  "/cgi/product_jqm2.pl": {
    /**
     * Add or Edit A Product 
     * @description If the barcode exists then you will be editing the existing product,
     * However if it doesn''t you will be creating a new product with that unique barcode,
     * and adding properties to the product.
     */
    post: operations["post-cgi-product_jqm2.pl"];
  };
  "/api/v2/search": {
    /**
     * Search for Products 
     * @description Search request allows you to get the nutritional data of products that match your search criteria.
     * It allows you create many custom APIs for your use case.
     * If the search query parameter has 2 possible values, they are seperated by a comma(,).
     * When filtering via a parameter that has different language codes like `fr`, `de` or `en`, specify the language code in the parameter name e.g `categories_tags_en`
     */
    get: operations["get-search"];
  };
  "/cgi/suggest.pl": {
    /**
     * Get Suggestions to Aid Adding/Editing Products 
     * @description For example , Dave is looking for packaging_shapes that contain the term "fe",
     * all packaging_shapes containing "fe" will be returned.
     * This is useful if you have a search in your application,
     * for a specific product field.
     */
    get: operations["get-cgi-suggest.pl"];
  };
  "/api/v2/product/{barcode}?fields=images": {
    /**
     * Get Images for a specific product by barcode (special case of get product) 
     * @description Images ensure the reliability of Open Food Facts data. It provides a primary source and proof of all the structured data. You may therefore want to display it along the structured information.
     */
    get: operations["get-api-v2-product-barcode-?fields=images"];
    parameters: {
      path: {
        barcode: string;
      };
    };
  };
}

export type webhooks = Record<string, never>;

export interface components {
  schemas: {
    Product: external["schemas/product.yaml"];
  };
  responses: never;
  parameters: {
    id: string;
    /** @description Barcode of the product */
    code: string;
    process_image: string;
    ocr_engine: string;
    imgid: string;
    angle: string;
    /**
     * @description The allowed values  used to sort/order the search results.
     * 
     * * `product_name` sorts on name
     * * `ecoscore_score`, `nova_score`, `nutriscore_score` rank on the [Eco-Score](https://world.openfoodfacts.org/eco-score-the-environmental-impact-of-food-products), [Nova](https://world.openfoodfacts.org/nova), or [Nutri-Score](https://world.openfoodfacts.org/nutriscore)
     * * `scans_n`, `unique_scans_n` and `popularity_key` are about product popularity: number of scans on unique scans, rank of product
     * * `created_t`, `last_modified_t`, are about creation and modification dates
     * * `nothing`, tells not to sort at all (because if you do not provide the sort_by argument we default to sorting on popularity (for food) or last modification date)
     */
    sort_by?: "product_name" | "last_modified_t" | "scans_n" | "unique_scans_n" | "created_t" | "completeness" | "popularity_key" | "nutriscore_score" | "nova_score" | "nothing" | "ecoscore_score";
    /**
     * @description The fields to be returned from the product object can also be limited.
     * If not specified, it returns the entire product object response.
     */
    fields?: string;
    tagtype?: string;
    term?: string;
  };
  requestBodies: never;
  headers: never;
  pathItems: never;
}
// ...
export interface operations {

  /**
   * Get information for a specific product by barcode 
   * @description A product can be fetched via its unique barcode.
   * It returns all the details of that product response.
   */
  "get-product-by-barcode": {
    parameters: {
      path: {
        /** @description The barcode of the product to be fetched */
        barcode: string;
      };
    };
    responses: {
      /** @description OK */
      200: {
        content: {
          "application/json": external["responses/get_product_by_barcode.yaml"];
        };
      };
    };
  };
// ...
  "get-search": {
    parameters: {
      query: {
        fields?: components["parameters"]["fields"];
        sort_by?: components["parameters"]["sort_by"];
      };
    };
    responses: {
      /** @description OK */
      200: {
        content: {
          "application/json": external["responses/search_for_products.yaml"];
        };
      };
    };
  };
// ...
}